Zoveel mogelijk van level 8 taak 9 gedaan

+++ ReversiApp/Controllers/SpelController.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
@@ -13,16 +14,27 @@ namespace ReversiApp.Controllers
     public class SpelController : Controller
     {
         private readonly ReversiContext _context;
+        private readonly UserManager<Speler> UserManager;
+        private readonly RoleManager<IdentityRole> RoleManager;
 
-        public SpelController(ReversiContext context)
+        public SpelController(ReversiContext context, UserManager<Speler> userManager, RoleManager<IdentityRole> roleManager)
         {
             _context = context;
+            UserManager = userManager;
+            RoleManager = roleManager;
         }
 
         // GET: Spel
         public async Task<IActionResult> Index()
         {
-            return View(await _context.Spel.ToListAsync());
+            Speler speler = await UserManager.FindByNameAsync(User.Identity.Name);
+            if(speler.SpelId != null)
+            {
+                return RedirectToAction(nameof(Game), new { ID = speler.SpelId });
+            }
+            var spellen = await _context.Spel.Include(s => s.Spelers).Where(s => s.Spelers.Count < 2).ToListAsync();
+            ViewBag.AantalSpellen = spellen.Count;
+            return View(spellen);
         }
 
         // GET: Spel/Details/5
@@ -43,6 +55,29 @@ namespace ReversiApp.Controllers
             return View(spel);
         }
 
+        // GET: Spel/Game/5
+        public async Task<IActionResult> Game(int? id)
+        {
+            if (id == null)
+            {
+                return NotFound();
+            }
+
+            var spel = await _context.Spel.Include(s => s.Spelers).FirstOrDefaultAsync(m => m.ID == id);
+            if (spel == null)
+            {
+                return NotFound();
+            }
+            foreach (var speler in spel.Spelers)
+            {
+                if (speler.Email == User.Identity.Name)
+                {
+                    return View(spel);
+                }
+            }
+            return RedirectToAction(nameof(Index));
+        }
+
         // GET: Spel/Create
         public IActionResult Create()
         {
@@ -60,7 +95,13 @@ namespace ReversiApp.Controllers
             {
                 _context.Add(spel);
                 await _context.SaveChangesAsync();
-                return RedirectToAction(nameof(Index));
+                var lastAddedGame = _context.Spel.ToList().LastOrDefault();
+
+                Speler speler = await UserManager.FindByNameAsync(User.Identity.Name);
+                speler.SpelId = lastAddedGame.ID;
+                var result = await UserManager.UpdateAsync(speler);
+
+                return RedirectToAction(nameof(Game), new { ID = lastAddedGame.ID });
             }
             return View(spel);
         }
@@ -116,6 +157,51 @@ namespace ReversiApp.Controllers
             return View(spel);
         }
 
+        // GET: Spel/Join/5
+        [HttpGet]
+        public async Task<IActionResult> Join(int? id)
+        {
+            if (id == null)
+            {
+                return NotFound();
+            }
+
+            var spel = await _context.Spel.FindAsync(id);
+            var spelersInSpel = _context.Speler.Where(s => s.SpelId == id).Count();
+
+            if (spel == null)
+            {
+                return NotFound();
+            }
+            else if (spelersInSpel >= 2)
+            {
+                return RedirectToAction(nameof(Index));
+            }
+            return View(spel);
+        }
+
+        // POST: Spel/Join/5
+        [HttpPost]
+        public async Task<IActionResult> Join(int id, [Bind("ID")] Spel spel)
+        {
+            if (id != spel.ID)
+            {
+                return NotFound();
+            }
+
+            Speler speler = await UserManager.FindByNameAsync(User.Identity.Name);
+            if (speler != null)
+            {
+                speler.SpelId = id;
+                var result = await UserManager.UpdateAsync(speler);
+                if (result.Succeeded)
+                {
+                    return RedirectToAction(nameof(Index));
+                }
+            }
+            return View(spel);
+        }
+
         // GET: Spel/Delete/5
         public async Task<IActionResult> Delete(int? id)
         {
@@ -139,9 +225,18 @@ namespace ReversiApp.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
+            Speler speler = await UserManager.FindByNameAsync(User.Identity.Name);
             var spel = await _context.Spel.FindAsync(id);
-            _context.Spel.Remove(spel);
-            await _context.SaveChangesAsync();
+            if (speler != null && spel != null)
+            {
+                speler.SpelId = null;
+                var result = await UserManager.UpdateAsync(speler);
+                if (result.Succeeded)
+                {
+                    _context.Spel.Remove(spel);
+                    await _context.SaveChangesAsync();
+                }
+            }
             return RedirectToAction(nameof(Index));
         }
